backend/requirements/base.txt
# Django
Django==5.1
djangorestframework==3.15.2
django-filter==24.3
django-environ==0.11.2

# Security
django-cors-headers==4.4.0
argon2-cffi==23.1.0

# Docs
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.2.1

# Images
cloudinary==1.38.0

# Utilities
pillow==10.3.0
django-import-export==4.0.7
Markdown==3.6
